requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
GitPython>=3.1,<4
openai>=1.0,<2
pydantic>=2.0,<3
pywebview>=5.0,<7
pyinstaller>=6.0,<7
